refactor(github): split PRPollingSource into domain/infra/value-object layers
PRPollingSource mixed four concerns in one 1226-line class. Separate them
along DDD lines so each is independently testable and the orchestrator only
wires them together:

- prCheckRunDomain.ts (pure domain): isCheckFailure, checkKey, ciTerminalStatus,
  computeNewCheckFailures, and planAnnotationCandidates — total transformations
  over plain check-run data, no I/O or subscription state.
- annotationFetchBudget.ts (value object): AnnotationFetchBudget, its exhausted
  error, the process-wide singleton, and the budget-window constants.
- checkRunsClient.ts (infrastructure): fetchAllCheckRuns / fetchAnnotations —
  the GitHub REST transport plus per-page ETag caching, taking a logger and the
  cache/budget by parameter. Exports a standalone CheckRunsCache type so the
  state can reference it without an import cycle.
- PRPollingSource.ts stays the orchestrator with thin delegators, preserving the
  same public/test-facing method surface and SubscriptionState shape.

Behavior is unchanged: typecheck passes and all 18 github polling tests stay
green without test edits.
